requirements.txt
streamlit
pytesseract
Pillow
pandas
pdf2image
opencv-python-headless
requests
xlsxwriter